requirements.txt
bottle==0.12.16
httplib2==0.12.1
lxml==4.3.3
PyYAML==5.1
uwsgi==2.0.18
